mod-ur: refactor 'common' folder to distinguish between file types
