+++ _ur/browser-client/@client.ts
@@ -8,7 +8,7 @@
 \*\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\\ * /////////////////////////////////////*/
 
 // note: cjs-style modules in 'common' can not be destructured on import
-import PROMPTS from '../common/prompts.js';
+import PROMPTS from '../common/util-prompts.js';
 const { makeStyleFormatter } = PROMPTS;
 import StateMgr from '../common/class-statemgr.ts';
