+++ _ur/common/util-text.js
