+++ _ur/node-server/class-urmodule.mts
@@ -12,7 +12,7 @@ import { ChildProcess } from 'node:child_process';
 
 /// CONSTANTS & DECLARATIONS //////////////////////////////////////////////////
 /// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
-import { makeTerminalOut } from '../common/prompts.js';
+import { makeTerminalOut } from '../common/util-prompts.js';
 const LOG = makeTerminalOut(' URMOD', 'TagYellow');
 const DBG = true;
