+++ _ur/node-server/files.mts
@@ -17,7 +17,7 @@
 import NDIR from 'node-dir';
 import FSE from 'fs-extra';
 import PATH from 'node:path';
-import PROMPT from '../common/prompts.js';
+import PROMPT from '../common/util-prompts.js';
 import * as url from 'url';
 
 /// CONSTANTS & DECLARATIONS //////////////////////////////////////////////////
