+++ _ur/npm-scripts/@build-web.cjs
@@ -15,7 +15,7 @@ const UR = require('@ursys/core');
 const { EnsureDir } = UR.FILES;
 // build-web can not use URSYS library because it's BUILDING it!
 // so we yoink the routines out of the source directly
-const PROMPTS = require('../common/prompts');
+const PROMPTS = require('../common/util-prompts');
 
 /// CONSTANTS AND DECLARATIONS ////////////////////////////////////////////////
 /// - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
